+++ routes/login.js
@@ -1,31 +1,7 @@
-let User = require("../controllers/userControllers");
+let LoginController = require("../controllers/loginControllers");
+const router = require("express").Router();
 const { body, validationResult } = require("express-validator");
-const httpStatusCodes = require("../common/httpCodes");
-const bcrypt = require("bcrypt");
 
-app.post("/", body("email").isEmail(), body("password").isLength({ min: 1 }), async (req, res) => {
-  try {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      throw new Error("Invalid email or password");
-    }
+router.post("/", body("email").isEmail(), body("password").isLength({ min: 1 }), LoginController.login);
 
-    let user = await User.findUserByEmail(req.body.email);
-
-    if (user === null) {
-      throw new Error("User/email not found");
-    } else {
-      let passwordIsOk = bcrypt.compareSync(req.body.password, user.password);
-
-      if (passwordIsOk) {
-        //Si todo está bien
-
-        res.json(user);
-      } else {
-        throw new Error("Invalid password");
-      }
-    }
-  } catch (error) {
-    res.status(httpStatusCodes.NOT_FOUND).send({ ok: false });
-  }
-});
+module.exports = router;
